Rename vector entry to ref, element to value
The naming conventions are being consolidated so that "ref" is used
everywhere as an abstract reference to an element in a container and
"value" is a value in that container. This makes vector match ring and
the new array_access template and the various sort algorithms.

Refs are also now always passed by value (except as an output parameter,
e.g. in emplace), and ref functions always take the container first like
any other container functions. This makes the array access interface
consistent between containers so it is easy to test them and easy to
wrap them in the array_access template.