Rename ref to entry for array containers
This essentially reverts a3719d9 for vector, and also renames entry in
ring and deque. Entry and ref need to be separate types for many
reasons, but the most immediate reason is fixing sort of arbitrary
containers. Using "entry" in the API for containers is also simpler and
easier to understand.