Change compare back to using refs
The compare template needs to operate on refs, not entries, because it
needs to be able to compare temporaries. (This is technically not
required for quick_sort but it is for insertion_sort among others.) The
new support for container_types and non-restrict lifecycle swap
simplifies both the compare and quick_sort code while making generalized
array sorting (e.g. deque_sort) possible.