Fix deque next/previous, first_bulk functions
The deque previously inserted elements at the front in reverse order to
maintain symmetry with the end, so you could use either end as the push
and the other as the pop for bulk values. This doesn't make sense in
practice because a) it was slow; b) insert_first_bulk() wasn't the
reverse of extract_first_bulk(), which was confusing; and c) I
eventually want to add insert_at_bulk() to deque, and it wouldn't make
sense for insert_first_bulk() to have different behaviour than
insert_at_bulk(0).

This fixes it to do the sensible thing. Inserting in bulk at the front
of the deque is the same as inserting in bulk at any other position in
the deque (when I eventually implement it.)

This also fixes some bugs in next/previous which allow the sort_deque
example to work properly.