Rename deque to pager
Pottery templates are named after the concrete algorithm they implement,
not its high-level interface. Deque means double-ended queue, and there
are many other Pottery templates that implement a double-ended queue:
ring, vector in double-ended mode, linked list (soon to be added), etc.

For lack of a better name, the Pottery analogue of std::deque is now
called "pager" since it is a ring of fixed-size pages.