Convert CLEANUP into proper boolean flag
Pottery usually avoids using #ifdef to check configuration. This helps
avoid bugs with macros. The NO_CLEANUP macro didn't follow this rule.

NO_CLEANUP is renamed to CLEANUP to avoid a double negative and is made
into a proper boolean flag. It defaults to 1 and must be defined to 0 to
disable automatic cleanup.