Move entry_t from container_types to array_access
This moves the type entry_t and the entry_ref() conversion from
container_types to array_access and to the various containers.

It didn't make sense for container_types to define entry_t; it's the
container that needs to define its own entry_t and conversion to ref.
This is why pager hasn't been converted to container_types yet.

This change is slightly breaking: the entry type and conversion on those
templates that use array_access (heap, open_hash_table and the array sort
algorithms) are now configured as ARRAY_ACCESS_ENTRY_TYPE and
ARRAY_ACCESS_ENTRY_REF. See the changes in the sort_pager example.

This will allow the various entry_value() functions to also be renamed
to entry_ref() making everything consistent between array_access and the
various container_types. This will be a more breaking change.