Add red_black_tree, tree_map
This is the first version of the red_black_tree and tree_map templates.
They do work but they are probably buggy, they need a lot of cleanup and
the red_black_tree API will probably change quite a bit.

The tree_map API probably won't change much because it's the same map
API as the others.