Rename pottery_qsort() functions
pottery_qsort_simple() is renamed to pottery_qsort() since it's the one
that's meant to be embedded in other projects.

The old pottery_qsort() is renamed to pottery_qsort_fast() and is
staying in examples/. It's not really meant to be used; it's mostly just
for demonstrating Pottery's speed and configurability.

A link from qsort/ to qsort_fast/ is sticking around to preserve old
URLs.